Move the file location of launchable.rb
