+++ tool/lib/launchable.rb
